Rename tile logo
