+++ client-react/src/components/TileImage/index.js
